backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
groq
python-dotenv
